app/app_requirements.txt
streamlit
langchain==0.2.16
langchain_experimental==0.0.65
langchain-openai==0.1.25
pydantic-settings==2.5.2
pandas
mysql-connector-python
pandasai
python-dotenv
openai